Implement multi-threading for different connections.
Reads and writes on the same connection are still serial due to the
serial nature of TCP. Reads to the same GCL from differen connections
are concurrent. Writes to the same GCL from different connections
are NOT concurrent.

So far, performance looks decent. There was about a 3x increase in
throughput with 8 connections to different logs. That will probably
go down when there are multiple writes to the same log.